+++ R/Hancock-package.R
@@ -1,4 +1,7 @@
 #' @keywords internal
-#' @importFrom methods setGeneric
-#' @import utils
+#' @import GSEABase
+#' @importFrom SummarizedExperiment colData colData<-
+#' @importFrom S4Vectors metadata metadata<-
+#' @importFrom methods setGeneric getPackageName
+#' @importFrom utils packageVersion
 "_PACKAGE"